Added part of remove instruction.
- Remove instruction is only missing breaking up molecules with no bonds
holding the atoms.

- The remaining functions to complete before bug fixing and testing the
simulation to make sure it is correct are

BreakUpMolecules()
Is_Molecule_OutOfBounds()
Do_Molecules_Overlap()

- I need to fix how molecules are outputted because they are done
incorrectly. When a output is triggered a single molecule goes out in a
packet. I currently have mine pack it into a packed molecule of
multiple. Something I have to figure out is which it decides to output
first. Also it freezes the Waldo on the output for X cycles until X
molecules are outputted (will cause a stall if it can't output).